+++ libs/components/src/lib/avatar/avatar.spec.ts
@@ -1,8 +1,9 @@
 import {elementUpdated, fixture, getBaseElement} from '@vivid-nx/shared';
 import {Connotation} from '../enums';
 import { Avatar } from './avatar';
-import '.';
 import type { Icon } from '../icon/icon';
+import '.';
+import { avatarDefinition } from './definition';
 
 const COMPONENT_TAG = 'vwc-avatar';
 
@@ -23,6 +24,7 @@ describe('vwc-avatar', () => {
 
 	describe('basic', () => {
 		it('should be initialized as a vwc-avatar', async () => {
+			expect(avatarDefinition()).toBeInstanceOf(FoundationElementRegistry);
 			expect(element).toBeInstanceOf(Avatar);
 		});
 	});
@@ -88,7 +90,7 @@ describe('vwc-avatar', () => {
 			iconElement = baseElement.querySelector('vwc-icon') as Icon;
 		});
 
-		it('should have the default icon', async () => { 
+		it('should have the default icon', async () => {
 			expect(iconElement?.getAttribute('name')).toEqual('user-line');
 		});