+++ libs/components/src/lib/card/card.spec.ts
@@ -1,7 +1,9 @@
 import {elementUpdated, fixture } from '@vivid-nx/shared';
+import { FoundationElementRegistry } from '@microsoft/fast-foundation';
 import { Icon } from '../icon/icon';
 import { Card } from './card';
 import '.';
+import { cardDefinition } from './definition';
 
 const COMPONENT_TAG = 'vwc-card';
 const ICON_SELECTOR = 'vwc-icon';
@@ -15,6 +17,7 @@ describe('vwc-card', () => {
 
 	describe('basic', () => {
 		it('should be initialized as a vwc-card', async () => {
+			expect(cardDefinition()).toBeInstanceOf(FoundationElementRegistry);
 			expect(element).toBeInstanceOf(Card);
 			expect(element.icon).toBeUndefined();
 		});