+++ libs/components/src/lib/fab/fab.spec.ts
@@ -1,6 +1,8 @@
 import { elementUpdated, fixture, getControlElement } from '@vivid-nx/shared';
+import { FoundationElementRegistry } from '@microsoft/fast-foundation';
 import { Fab, FabConnotation } from './fab';
 import '.';
+import { fabDefinition } from './definition';
 
 const COMPONENT_TAG = 'vwc-fab';
 
@@ -13,6 +15,7 @@ describe('vwc-fab', () => {
 
 	describe('basic', () => {
 		it('should be initialized as a vwc-fab', async () => {
+			expect(fabDefinition()).toBeInstanceOf(FoundationElementRegistry);
 			expect(element).toBeInstanceOf(Fab);
 			expect(element.label).toBeUndefined();
 			expect(element.icon).toBeUndefined();