+++ libs/components/src/lib/menu-item/menu-item.spec.ts
@@ -1,10 +1,11 @@
 import { elementUpdated, fixture } from '@vivid-nx/shared';
 import '.';
-import { MenuItemRole } from '@microsoft/fast-foundation';
+import { FoundationElementRegistry, MenuItemRole } from '@microsoft/fast-foundation';
 import { keyEnter, keySpace } from '@microsoft/fast-web-utilities';
 import { fireEvent } from '@testing-library/dom';
 import type { Icon } from '../icon/icon';
 import { MenuItem } from './menu-item';
+import { menuItemDefinition } from './definition';
 
 
 const COMPONENT_TAG = 'vwc-menu-item';
@@ -25,6 +26,7 @@ describe('vwc-menu-item', () => {
 
 	describe('basic', () => {
 		it('should be initialized as a vwc-menu-item', async () => {
+			expect(menuItemDefinition()).toBeInstanceOf(FoundationElementRegistry);
 			expect(element).toBeInstanceOf(MenuItem);
 			expect(element.text).toEqual(undefined);
 			expect(element.role).toEqual('menuitem');