+++ libs/components/src/lib/progress-ring/definition.ts
@@ -1,5 +1,5 @@
 import type {FoundationElementDefinition} from '@microsoft/fast-foundation';
-import { registerFactorial } from '../../shared/design-system';
+import { registerFactory } from '../../shared/design-system';
 import styles from './progress-ring.scss';
 
 import {ProgressRing} from './progress-ring';
@@ -7,21 +7,22 @@ import {ProgressRingTemplate as template} from './progress-ring.template';
 
 /**
  * The progress-ring element.
- *
- * @internal
  */
-export const progressRing =
+export const progressRingDefinition =
 	ProgressRing.compose<FoundationElementDefinition>({
 		baseName: 'progress-ring',
 		template: template as any,
 		styles,
-	})();
+	});
 
-export const progressRingElements = [progressRing];
+/**
+ * @internal
+ */
+export const progressRingRegistries = [progressRingDefinition()];
 
 /**
  * Registers the progress-ring  elements with the design system.
  *
  * @param prefix - the prefix to use for the component name
  */
-export const registerProgressRing = registerFactorial(progressRingElements);
+export const registerProgressRing = registerFactory(progressRingRegistries);